backend/requirements.txt
fastapi[standard]
openai
faiss-cpu
numpy
pydantic-settings
python-dotenv
pytest
ruff
mypy
bandit
